feat(Rendering): add coordinate systems for props
This MR adds the noption of coordinate systems to vtkProp.
It currently adds WORLD as the default and DISPLAY as a
second option. DISPLAY coordinates are in pixels with 0,0 being
the lower left corner of the viewport and Z ranging from -1
near to 1 far. These coordinate systems are intended to be
implemented in the GPU.

This MR adds Actor2D and PolyDataMapper2D to WebGPU taking
advantage of the new coordinate systems.

This MR updates the ScalarBarActor to produce data in
DISPLAY coordinates. The old code used world coordinates
which meant that every time the camera changed the scalar
bar actor had to recompute it's data which is very expensive
resulting in hundreds of VBO's being created every second.
By using DISPLAY coordinate the data does not change with
camera changes, only with viewport changes or changes to
the source data/scalar bar actor itself which is far less
common.

With this MR the scalar bar actor functions on both WebGPU and
OpenGL and the Actor2D example works on both WebGPU and OpenGL.